+++ src/scripts/utils/constants.ts
@@ -6,4 +6,4 @@ export const CLI_OPTION_INDENT = '  ';
 /**
  * An amount to indent presented progress items in the CLI.
  */
-export const CLI_PROGRESS_ITEM_INDENT = '   ';
+export const CLI_PROGRESS_ITEM_INDENT = '  ';