+++ src/configs/eslint/index.ts
@@ -1,8 +1,8 @@
 import prettierConfig from 'eslint-config-prettier/flat';
 
 import { nodeConfig } from './node.config.js';
+import { reactConfig } from './react.config.js';
 import { reactNativeConfig } from './reactNative.config.js';
-import { reactNextConfig } from './reactNext.config.js';
 import { reactViteConfig } from './reactVite.config.js';
 import { ProjectType } from '../../types/index.js';
 
@@ -23,7 +23,7 @@ export const eslintConfigs: Record<ProjectType, Linter.Config[]> = {
      * CodeFish Studio ESLint configuration for React (Next.js) + TypeScript projects
      */
     reactNext: [
-        ...reactNextConfig,
+        ...reactConfig,
 
         // Prettier must come last to override conflicting rules
         prettierConfig,