+++ src/configs/eslint/react.config.ts
@@ -3,16 +3,7 @@ import importPlugin from 'eslint-plugin-import';
 import reactPlugin from 'eslint-plugin-react';
 import * as reactHooks from 'eslint-plugin-react-hooks';
 
-import { baseConfig, dependencies as baseDependencies } from './base.config.js';
-
-/**
- * Dependencies required for the React Universal ESLint configuration
- */
-export const dependencies = {
-    ...baseDependencies,
-    'eslint-plugin-react': '^7.37',
-    'eslint-plugin-react-hooks': '6.0.0-rc.1',
-} as const;
+import { baseConfig } from './base.config.js';
 
 const reactPluginConfig = {
     ...reactPlugin.configs.flat.recommended,