+++ src/configs/eslint/reactNative.config.ts
@@ -2,16 +2,7 @@ import { fixupPluginRules } from '@eslint/compat';
 import { defineConfig } from 'eslint/config';
 import reactNativePlugin from 'eslint-plugin-react-native';
 
-import { reactConfig, dependencies as reactDependencies } from './react.config.js';
-
-/**
- * Dependencies required for the React Native ESLint configuration
- */
-export const dependencies = {
-    ...reactDependencies,
-    '@eslint/compat': '^1',
-    'eslint-plugin-react-native': '^5',
-} as const;
+import { reactConfig } from './react.config.js';
 
 /**
  * React Native project ESLint configuration.